+++ editors/vscode/src/properties_client.ts
@@ -4,7 +4,7 @@
 import {
     PropertyQuery,
     SetBindingResponse,
-} from "../../../tools/online_editor/src/shared/properties";
+} from "../../../tools/slintpad/src/shared/properties";
 
 import {
     OptionalVersionedTextDocumentIdentifier,